Fix bit shifting bug

+++ Source/com/drew/lang/ByteUtil.java
@@ -36,11 +36,11 @@ public class ByteUtil
     {
         if (bigEndian) {
             return
-                ((buffer[offset    ] & 0xFF) << 56) |
-                ((buffer[offset + 1] & 0xFF) << 48) |
-                ((buffer[offset + 2] & 0xFF) << 40) |
-                ((buffer[offset + 3] & 0xFF) << 32) |
-                ((buffer[offset + 4] & 0xFF) << 24) |
+                ((long) (buffer[offset    ] & 0xFF) << 56) |
+                ((long) (buffer[offset + 1] & 0xFF) << 48) |
+                ((long) (buffer[offset + 2] & 0xFF) << 40) |
+                ((long) (buffer[offset + 3] & 0xFF) << 32) |
+                ((long) (buffer[offset + 4] & 0xFF) << 24) |
                 ((buffer[offset + 5] & 0xFF) << 16) |
                 ((buffer[offset + 6] & 0xFF) <<  8) |
                 ((buffer[offset + 7] & 0xFF));
@@ -50,11 +50,11 @@ public class ByteUtil
                 ((buffer[offset    ] & 0xFF)      ) |
                 ((buffer[offset + 1] & 0xFF) <<  8) |
                 ((buffer[offset + 2] & 0xFF) << 16) |
-                ((buffer[offset + 3] & 0xFF) << 24) |
-                ((buffer[offset + 4] & 0xFF) << 32) |
-                ((buffer[offset + 5] & 0xFF) << 40) |
-                ((buffer[offset + 6] & 0xFF) << 48) |
-                ((buffer[offset + 7] & 0xFF) << 56);
+                ((long) (buffer[offset + 3] & 0xFF) << 24) |
+                ((long) (buffer[offset + 4] & 0xFF) << 32) |
+                ((long) (buffer[offset + 5] & 0xFF) << 40) |
+                ((long) (buffer[offset + 6] & 0xFF) << 48) |
+                ((long) (buffer[offset + 7] & 0xFF) << 56);
         }
     }
 }
